requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
jinja2==3.1.6
python-multipart==0.0.21
pyjwt==2.10.1
lz4==4.4.5
pydantic==2.10.4
pydantic-settings==2.7.1
